pyproject.toml
[project]
name = "orochi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.13.2"
dependencies = [
    "argon2-cffi==25.1.0",
    "beautifulsoup4==4.13.4",
    "bokeh>=3.7.3",
    "capstone==5.0.6",
    "channels-redis==4.3.0",
    "channels[daphne]==4.3.1",
    "clamdpy==0.1.0.post1",
    "cloudpickle==3.1.1",
    "crispy-bootstrap5==2025.6",
    "daphne==4.2.1",
    "dask==2025.7.0",
    "distributed==2025.7.0",
    "django==5.2.5",
    "django-admin-list-filter-dropdown==1.0.3",
    "django-admin-multiple-choice-list-filter==0.1.1",
    "django-allauth[mfa]==65.10.0",
    "django-auth-ldap==5.2.0",
    "django-colorfield==0.14.0",
    "django-cors-headers==4.7.0",
    "django-crispy-forms==2.4",
    "django-environ==0.12.0",
    "django-extra-settings==0.14.0",
    "django-file-form==3.9.0",
    "django-guardian==3.0.3",
    "django-import-export==4.3.9",
    "django-json-widget==2.0.3",
    "django-ninja==1.4.3",
    "django-redis==6.0.0",
    "django-widget-tweaks==1.5.0",
    "geoip2==5.1.0",
    "gitpython==3.1.45",
    "gunicorn==23.0.0",
    "hiredis==3.2.1",
    "httpx>=0.28.1",
    "jsonschema==4.25.0",
    "jupyter-server-proxy>=4.4.0",
    "lz4==4.4.4",
    "marko==2.2.0",
    "msgpack==1.1.1",
    "numpy==2.3.2",
    "pandas==2.3.1",
    "pefile==2024.8.26",
    "pillow==11.3.0",
    "plotly==6.2.0",
    "psycopg2-binary==2.9.10",
    "pyarrow==21.0.0",
    "pycryptodome==3.23.0",
    "pymisp==2.5.17",
    "python-debian==1.0.1",
    "python-ldap==3.4.4",
    "python-magic==0.4.27",
    "pytz==2025.2",
    "redis==6.4.0",
    "regipy[full]==5.2.0",
    "rpmfile==2.1.0",
    "toolz==1.0.0",
    "tornado==6.5.2",
    "volatility3",
    "vt-py==0.21.0",
    "whitenoise==6.9.0",
    "yara-x==1.5.0",
]

[tool.uv.sources]
volatility3 = { git = "https://github.com/volatilityfoundation/volatility3.git", rev = "1ebb82a0c0ec00c6e54f2d5f731de18520367041" }

[dependency-groups]
dev = [
    "black==25.1.0",
    "coverage==7.10.2",
    "django-coverage-plugin==3.1.1",
    "django-debug-toolbar==6.0.0",
    "django-extensions==4.1",
    "django-stubs==5.2.2",
    "factory-boy==3.3.3",
    "flake8==7.3.0",
    "flake8-isort==6.1.2",
    "mypy==1.17.1",
    "pre-commit==4.2.0",
    "pylint-django==2.6.1",
    "pytest==8.4.1",
    "pytest-django==4.11.1",
    "pytest-sugar==1.0.0",
]
